chore: update package versions
