+++ packages/imready/package.json
@@ -67,7 +67,7 @@
 		"rollup-plugin-uglify": "^6.0.2",
 		"sinon": "^7.3.0",
 		"tslib": "^2.0.3",
-		"typescript": "~4.1"
+		"typescript": "~4.8.4"
 	},
 	"dependencies": {
 		"@cfcs/core": "^0.1.0",
