+++ packages/ngx-imready/projects/ngx-imready/package.json
@@ -2,7 +2,7 @@
   "name": "@egjs/ngx-imready",
   "version": "1.1.1",
   "dependencies": {
-    "@cfcs/angular": "^0.1.0",
-    "@egjs/imready": "~1.3.1"
+    "@cfcs/angular": "^0.1.1",
+    "@egjs/imready": "~1.4.0"
   }
 }
