+++ packages/react-imready/package.json
@@ -29,7 +29,7 @@
     "react-dom": "^17.0.1",
     "react-scripts": "4.0.3",
     "rollup": "^2.33.3",
-    "typescript": "~4.1"
+    "typescript": "~4.8.4"
   },
   "scripts": {
     "start": "SKIP_PREFLIGHT_CHECK=true react-scripts start",
