+++ packages/svelte-imready/package.json
@@ -34,7 +34,7 @@
   "scripts": {
     "dev": "rollup -c -w",
     "start": "sirv public --no-clear",
-    "build": "rollup -c rollup.build.config.js && npm run declaration",
+    "build": "rollup -c && npm run declaration",
     "declaration": "rm -rf declaration && tsc -p tsconfig.declaration.json"
   },
   "devDependencies": {
@@ -59,7 +59,7 @@
     "svelte-preprocess": "^4.6.1",
     "ts-jest": "^26.4.4",
     "tslib": "^1.10.0",
-    "typescript": "~4.1"
+    "typescript": "~4.8.4"
   },
   "dependencies": {
     "@cfcs/svelte": "^0.1.0",
