+++ packages/vue-imready/package.json
@@ -48,7 +48,7 @@
     "eslint": "^6.7.2",
     "eslint-plugin-vue": "^7.0.0-0",
     "print-sizes": "^0.1.0",
-    "typescript": "~4.1",
+    "typescript": "~4.8.4",
     "vue": "^3.0.0"
   }
 }
