+++ packages/vue2-imready/package.json
@@ -50,7 +50,7 @@
     "print-sizes": "^0.1.0",
     "rollup-plugin-vue": "^5.1.9",
     "tslib": "^2.3.1",
-    "typescript": "~4.1",
+    "typescript": "~4.8.4",
     "vue": "^2.6.12",
     "vue-template-compiler": "^2.6.11"
   },
